Finally done sucessfully

+++ InfowareTask/src/main/java/com/infoware/controllers/EmployeeController.java
@@ -1,3 +1,4 @@
+//This is controller class
 package com.infoware.controllers;
 
 import java.util.List;
@@ -25,43 +26,43 @@ public class EmployeeController{
 	@Autowired
 	EmployeeService service ; 
 
+	// to handle home page
 	@GetMapping("/")
 	public String showHomePage() {
 		return "home";
 	}
 	
+	// to show register form
 	@GetMapping("/register")
-	public String showRegisterForm() {
-		
+	public String showRegisterForm() {	
 		return "add";
 	}
 	
+	// to reguster the employee
 	@PostMapping("/register")
 	public String registerEmployee( @ModelAttribute("emp") Employee emp , RedirectAttributes attr) {
+		if(emp==null)
+			return "redirect:/error";
 		String result =null;
-		try {
-			
+		try {	
 		 result = service.registerEmp(emp);
-         attr.addFlashAttribute("result1","Employee registered successfully!");
+         attr.addFlashAttribute("result1",result);
 		 System.out.println(result);
-	}
+	     }
 		catch(Exception e)
 		{
 			System.out.println("EmployeeController.registerEmployee() "+e.getMessage());
 		}
-	return "redirect:/register";
+	return "redirect:/register";   // redirecting  to register form
 	}
 	
 	
+	// to show all registered employees
 	@GetMapping("/all_employee")
 	public String getAllEmployee(Map<String,List<Employee>> map) {
 		List<Employee> empList =null;
 		try {
-			
 			empList = service.getAllEmployee();
-			if(empList==null) {
-				
-			}
 			map.put("empList", empList);
 			for(Employee e : empList) {
 				System.out.println(e.getPhoneNo()+ "id--> "+e.getId());
@@ -73,16 +74,13 @@ public class EmployeeController{
 		}
 		return "fetchAllEmployee";
  }
+	
 	// show employee for delete
 	@GetMapping("/show_delete")
 	public String showEmpForDelete(Map<String,List<Employee>> map) {
 		List<Employee> empList =null;
 		try {
-			
 			empList = service.getAllEmployee();
-			if(empList==null) {
-				
-			}
 			map.put("empList", empList);
 		}
 		catch(Exception e)
@@ -92,10 +90,11 @@ public class EmployeeController{
 		return "delete";
 	}
 	
-	
 	// delete employee by id
 	@GetMapping("/delete")
 	public String deleteEmp(@RequestParam("id") Integer id, RedirectAttributes attr) {
+		if(id==null)
+			return "redirect:/error";
 		try {
 			service.deleteEmpById(id);
 			attr.addFlashAttribute("result3","Employee Deleted Successfully!!" );
@@ -106,6 +105,21 @@ public class EmployeeController{
 		}
 		return "redirect:/show_delete";
 	}
+	
+	// show all employee for update
+		@GetMapping("/show_update")
+			public String showEmpForUpdate(Map<String,List<Employee>> map) {
+				List<Employee> empList =null;
+				try {
+					empList = service.getAllEmployee();
+					map.put("empList", empList);
+				}
+				catch(Exception e)
+				{
+					System.out.println("EmployeeController.getAllEmployee() "+e.getMessage());
+				}
+				return "show_update";	
+			}
 	
 	//show updation form
 	@GetMapping("/update_emp")
@@ -122,16 +136,16 @@ public class EmployeeController{
 		return "update";
 	}
 	
+	// updating employee details
 	@PostMapping("/update")
 	public String Update( @ModelAttribute("emp") Employee emp, RedirectAttributes attr)
-	{
+	{   if(emp==null)
+		return "redirect:/error";
 		String result =null;
 		try {
-			
 		 result = service.updateEmployee(emp);
 		 attr.addFlashAttribute("result2",result);
-
-	}
+	   }
 		catch(Exception e)
 		{
 			System.out.println("EmployeeController.registerEmployee() "+e.getMessage());
@@ -139,37 +153,24 @@ public class EmployeeController{
 	return "redirect:/show_update";
 	}
 	
-	
-	@GetMapping("/show_update")
-		public String showEmpForUpdate(Map<String,List<Employee>> map) {
-			List<Employee> empList =null;
-			try {
-				
-				empList = service.getAllEmployee();
-				if(empList==null) {
-					
-				}
-				map.put("empList", empList);
-			}
-			catch(Exception e)
-			{
-				System.out.println("EmployeeController.getAllEmployee() "+e.getMessage());
-			}
-			return "show_update";
-			
-		}
-	
+	// to show pagination view
 	@GetMapping("/get_pagination")
 	public  String  showEmployeeReport(@PageableDefault(page=0,size=5,sort = "id",direction =Sort.Direction.ASC ) Pageable pageable,
 			                                                      Map<String,Object> map) {		
-		  //use serivce
+		try{
 		Page<Employee> page=service.getEmployeesPageData(pageable);
 		// put the results in model attributes
 		map.put("empsData",page);
-		//return LVN
-		return "pagination_view";
+		}
+	catch(Exception e) {
+		System.out.println("EmployeeController.showEmployeeReport()");
+		return "redirect:/error";
+	}
+	return "pagination_view";
 	}//method
 	
+	
+	// to show error page
 	@GetMapping("/error")
 	public String errorCode() {
 		return "error";
